[CHORE] : 파일 및 코드 수정
